core/api.py
from rest_framework.decorators import api_view
from rest_framework.response import Response


@api_view(['GET'])
def health(request):
    return Response({"message": "The API is healthy!"}, status=200)
