pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "psycopg>=3.3.2",
]

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "conf.settings"
python_files = ["test_*.py"]
